[Travis] Increase functional tests reserved time
Functional tests run in travis are given 1700 seconds. We are seeing more and more builds where it's not enough. So change the compile phase timeout to 800 seconds in order to have 2200 seconds for tests or cache the built binaries and restart with a full 3000 seconds for tests.
Github-Pull: #1157
Rebased-From: 63af787d055889d20f6a43ab1ff8ad1506f8c7e4