[Qt] Initialize isLoading to false for CS view
Loading of the cold staking delegations list checks to see if `isLoading`
is true, and exit's out if it is. this check is used to prevent double
loading, but on some systems `isLoading` is initialized as `true`, so
the loading of CS delegations will never occur.

This changes to initialization to be explicitly `false`.

Github-Pull: #1253
Rebased-From: d8378031521fe7af44c7d9c82115090e522d8245